+++ tests.py
@@ -1,44 +1,12 @@
 import pytest
-from main import BooksCollector
 
-
-@pytest.fixture
-def collector():
-    collector = BooksCollector()
-    return collector
-
-@pytest.fixture
-def collector_with_books_and_genres():
-    collector = BooksCollector()
-    collector.add_new_book('AHAHA')
-    collector.set_book_genre('AHAHA', 'Ужасы')
-    collector.add_new_book('OHOHO')
-    collector.set_book_genre('OHOHO', 'Комедии')
-    return collector
 
 # класс TestBooksCollector объединяет набор тестов, которыми мы покрываем наше приложение BooksCollector
 # обязательно указывать префикс Test
 class TestBooksCollector:
 
-    # # пример теста:
-    # # обязательно указывать префикс test_
-    # # дальше идет название метода, который тестируем add_new_book_
-    # # затем, что тестируем add_two_books - добавление двух книг
-    # def test_add_new_book_add_two_books(self):
-    #     # создаем экземпляр (объект) класса BooksCollector
-    #     collector = BooksCollector()
-    #
-    #     # добавляем две книги
-    #     collector.add_new_book('Гордость и предубеждение и зомби')
-    #     collector.add_new_book('Что делать, если ваш кот хочет вас убить')
-    #
-    #     # проверяем, что добавилось именно две
-    #     # словарь books_rating, который нам возвращает метод get_books_rating, имеет длину 2
-    #     assert len(collector.get_books_rating()) == 2
-
     # напиши свои тесты ниже
     # чтобы тесты были независимыми в каждом из них создавай отдельный экземпляр класса BooksCollector()
-
 
     # 1 дефолтный список жанров создается правильно
     def test_default_value_genres_true(self, collector):
